Further constrain geo search west/east of query location
The number of cells to the west and east of the cell containing the lookup point can be computed conservatively by measuring the west/east distance along the shortest edge of a grid cell.
The lookup region now resembles a rectangle distance-wise which provides a better fit than whole bands when there are several densely populated areas on similar latitudes.